leavescounter/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
scipy>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
trimesh>=4.0.0
requests>=2.31.0
